Zoom out the map
This also uses the Map’s options instead of an animateCamera. The reason is that the animation is not seen anyway when we’re opening the screen.

I’m keeping the reference to the `mapController` in case we want to do something fun with it later. (I have an idea.)